feature cxx17: support string_view print
