+++ test/logger_tests.hpp
@@ -51,4 +51,30 @@ public:
 
 #endif
 
+#if !defined(IUTEST_USE_GTEST)
+
+class LogChecker
+{
+    TestLogger printer_logger;
+    ::std::string m_str;
+public:
+    explicit LogChecker(const char* str) : m_str(str)
+    {
+        ::iutest::detail::iuConsole::SetLogger(&printer_logger);
+    }
+    ~LogChecker(void)
+    {
+        ::iutest::detail::iuConsole::SetLogger(NULL);
+        IUTEST_EXPECT_STRIN(m_str.c_str(), printer_logger.c_str());
+        printer_logger.clear();
+    }
+};
+#else
+class LogChecker
+{
+public:
+    explicit LogChecker(const char*) {}
+};
+#endif
+
 #endif
